Rename base_player.py to player.py
