+++ README.md
@@ -5,7 +5,7 @@
 ### Overview 
 This repository contains a python implementation of the [Wave Function Collapse](https://github.com/mxgmn/WaveFunctionCollapse) algorithm in [Python](https://www.python.org/) with the tile implementation presented on [this video](https://www.youtube.com/watch?v=qRtrj6Pua2A).  
 
-This mini project can best be described as a simplistic terminal based battle simulator. Where the battling entities are inherited instances of the included `base_player.py` containing an implementation of the evolution function:
+This mini project can best be described as a simplistic terminal based battle simulator. Where the battling entities are inherited instances of the included `player.py` containing an implementation of the evolution function:
 ```Python
 def evolve(self, curr_attack: int, curr_health: int, curr_movement: int, level: int) -> Evolution:
         """
